Login/Course Code/Admin Functionality Changes
When the student user logs in the system tests if the student table has been populated. If it has not, it will populate it (the system will only populate the student table on the first student user login). The system was also changed to check during login if the user had enrolled in a course yet. If they had not, the system will redirect them to an add_class page. Once they enter a class code, the enrollment table will be populated to connect the user to the course. The progress table still needs to be initialized when the student adds a course. Also, the admin view for several of the course and game views was updated to show more detailed information.