Progress Table Population
After the user adds a class code, the system will now create the progress table records for the student, and the problems in the packs. Also, if a user attempts to login with an account that doesn't exist, it will now return them back to the login page, instead of redirecting them to Google (This was previously used for testing purposes).